requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
pillow>=10.0.0
requests>=2.31.0
jupyter>=1.0.0

# Streamlit for web UI
streamlit>=1.28.0

# Azure AI Search
azure-search-documents>=11.4.0
azure-identity>=1.15.0

# Text processing
langchain-text-splitters>=0.0.1
